+++ management/src/main/java/com/foodielog/management/member/controller/dto/BlockProcessedParam.java
@@ -1,0 +1,13 @@
+package com.foodielog.management.member.controller.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class BlockProcessedParam {
+
+	private Long userId;
+
+	private String reason;
+}
