+++ management/src/main/java/com/foodielog/management/member/service/MemberService.java
@@ -1,9 +1,17 @@
 package com.foodielog.management.member.service;
 
-import com.foodielog.management.member.dto.request.BlockReq;
-import com.foodielog.management.member.dto.response.BadgeApplyListResp;
-import com.foodielog.management.member.dto.response.MemberListResp;
-import com.foodielog.management.member.dto.response.WithdrawListResp;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.foodielog.management.member.controller.dto.BlockProcessedParam;
+import com.foodielog.management.member.service.dto.BadgeApplyListResp;
+import com.foodielog.management.member.service.dto.MemberListResp;
+import com.foodielog.management.member.service.dto.WithdrawListResp;
 import com.foodielog.server._core.error.exception.Exception400;
 import com.foodielog.server._core.error.exception.Exception404;
 import com.foodielog.server._core.smtp.MailService;
@@ -19,153 +27,148 @@ import com.foodielog.server.feed.repository.FeedRepository;
 import com.foodielog.server.feed.type.ContentStatus;
 import com.foodielog.server.reply.entity.Reply;
 import com.foodielog.server.reply.repository.ReplyRepository;
-import com.foodielog.server.report.entity.Report;
 import com.foodielog.server.report.repository.ReportRepository;
 import com.foodielog.server.user.entity.User;
 import com.foodielog.server.user.repository.FollowRepository;
 import com.foodielog.server.user.repository.UserRepository;
 import com.foodielog.server.user.type.Flag;
 import com.foodielog.server.user.type.UserStatus;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.stream.Collectors;
+import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @Service
 public class MemberService {
 
-    private final WithdrawUserRepository withdrawUserRepository;
-    private final BadgeApplyRepository badgeApplyRepository;
-    private final FeedRepository feedRepository;
-    private final ReplyRepository replyRepository;
-    private final FollowRepository followRepository;
-    private final UserRepository userRepository;
-    private final ReportRepository reportRepository;
-    private final BlockUserRepository blockUserRepository;
+	private final WithdrawUserRepository withdrawUserRepository;
+	private final BadgeApplyRepository badgeApplyRepository;
+	private final FeedRepository feedRepository;
+	private final ReplyRepository replyRepository;
+	private final FollowRepository followRepository;
+	private final UserRepository userRepository;
+	private final ReportRepository reportRepository;
+	private final BlockUserRepository blockUserRepository;
 
-    private final MailService mailService;
+	private final MailService mailService;
 
-    @Transactional(readOnly = true)
-    public WithdrawListResp getWithdrawList(String nickName, Flag badge, Pageable pageable) {
-        List<WithdrawUser> withdrawUserList = withdrawUserRepository.findByFlag(pageable, nickName, badge);
+	@Transactional(readOnly = true)
+	public WithdrawListResp getWithdrawList(String nickName, Flag badge, Pageable pageable) {
+		List<WithdrawUser> withdrawUserList = withdrawUserRepository.findByFlag(pageable, nickName, badge);
 
-        List<WithdrawListResp.WithdrawMemberDTO> withdrawMemberDTOS = withdrawUserList.stream()
-                .map(WithdrawListResp.WithdrawMemberDTO::new)
-                .collect(Collectors.toList());
+		List<WithdrawListResp.WithdrawMemberDTO> withdrawMemberDTOS = withdrawUserList.stream()
+			.map(WithdrawListResp.WithdrawMemberDTO::new)
+			.collect(Collectors.toList());
 
-        return new WithdrawListResp(withdrawMemberDTOS);
-    }
+		return new WithdrawListResp(withdrawMemberDTOS);
+	}
 
-    @Transactional
-    public void restoreMember(Long withdrawId) {
-        WithdrawUser withdrawUser = withdrawUserRepository.findById(withdrawId)
-                .orElseThrow(() -> new Exception404("해당 회원을 찾을 수 없습니다."));
+	@Transactional
+	public void restoreMember(Long withdrawId) {
+		WithdrawUser withdrawUser = withdrawUserRepository.findById(withdrawId)
+			.orElseThrow(() -> new Exception404("해당 회원을 찾을 수 없습니다."));
 
-        User user = withdrawUser.getUser();
-        user.restore();
-        withdrawUserRepository.delete(withdrawUser);
+		User user = withdrawUser.getUser();
+		user.restore();
+		withdrawUserRepository.delete(withdrawUser);
 
-        List<Feed> feedList = feedRepository.findByUserIdAndStatus(user.getId(), ContentStatus.DELETED);
-        feedList.forEach(Feed::restoreFeed);
+		List<Feed> feedList = feedRepository.findByUserIdAndStatus(user.getId(), ContentStatus.DELETED);
+		feedList.forEach(Feed::restoreFeed);
 
-        List<Reply> replyList = replyRepository.findByUserIdAndStatus(user.getId(), ContentStatus.DELETED);
-        replyList.forEach(Reply::restoreReply);
-    }
+		List<Reply> replyList = replyRepository.findByUserIdAndStatus(user.getId(), ContentStatus.DELETED);
+		replyList.forEach(Reply::restoreReply);
+	}
 
-    @Transactional(readOnly = true)
-    public BadgeApplyListResp getBadgeApplyList(String nickName, ProcessedStatus processedStatus, Pageable pageable) {
-        List<BadgeApply> badgeApplyList = badgeApplyRepository.findByStatus(pageable, nickName, processedStatus);
+	@Transactional(readOnly = true)
+	public BadgeApplyListResp getBadgeApplyList(String nickName, ProcessedStatus processedStatus, Pageable pageable) {
+		List<BadgeApply> badgeApplyList = badgeApplyRepository.findByStatus(pageable, nickName, processedStatus);
 
-        List<BadgeApplyListResp.BadgeApplyMemberDTO> badgeApplyMemberDTOS = new ArrayList<>();
+		List<BadgeApplyListResp.BadgeApplyMemberDTO> badgeApplyMemberDTOS = new ArrayList<>();
 
-        for (BadgeApply badgeApply : badgeApplyList) {
-            User user = badgeApply.getUser();
+		for (BadgeApply badgeApply : badgeApplyList) {
+			User user = badgeApply.getUser();
 
-            Long feedCount = feedRepository.countByUserAndStatus(user, ContentStatus.NORMAL);
-            Long replyCount = replyRepository.countByUserAndStatus(user, ContentStatus.NORMAL);
-            Long followerCount = followRepository.countByFollowedId(user);
+			Long feedCount = feedRepository.countByUserAndStatus(user, ContentStatus.NORMAL);
+			Long replyCount = replyRepository.countByUserAndStatus(user, ContentStatus.NORMAL);
+			Long followerCount = followRepository.countByFollowedId(user);
 
-            BadgeApplyListResp.BadgeApplyMemberDTO badgeApplyMemberDTO = new BadgeApplyListResp.BadgeApplyMemberDTO(badgeApply, feedCount, replyCount, followerCount);
-            badgeApplyMemberDTOS.add(badgeApplyMemberDTO);
-        }
+			BadgeApplyListResp.BadgeApplyMemberDTO badgeApplyMemberDTO = new BadgeApplyListResp.BadgeApplyMemberDTO(
+				badgeApply, feedCount, replyCount, followerCount);
+			badgeApplyMemberDTOS.add(badgeApplyMemberDTO);
+		}
 
-        return new BadgeApplyListResp(badgeApplyMemberDTOS);
-    }
+		return new BadgeApplyListResp(badgeApplyMemberDTOS);
+	}
 
-    @Transactional
-    public void badgeProcessed(Long badgeApplyId, String process) {
-        BadgeApply badgeApply = badgeApplyRepository.findByIdAndStatus(badgeApplyId, ProcessedStatus.UNPROCESSED)
-                .orElseThrow(() -> new Exception404("해당 신청이 존재하지 않습니다."));
+	@Transactional
+	public void badgeProcessed(Long badgeApplyId, String process) {
+		BadgeApply badgeApply = badgeApplyRepository.findByIdAndStatus(badgeApplyId, ProcessedStatus.UNPROCESSED)
+			.orElseThrow(() -> new Exception404("해당 신청이 존재하지 않습니다."));
 
-        User user = badgeApply.getUser();
-        validateUserStatus(user);
+		User user = badgeApply.getUser();
+		validateUserStatus(user);
 
-        switch (process) {
-            case "rejected":
-                badgeApply.rejectBadge();
-                break;
-            case "approved":
-                badgeApply.approveBadge();
-                user.badgeApproved();
-                break;
-            default:
-                throw new Exception400(process, "잘못된 요청값입니다.");
-        }
-    }
+		switch (process) {
+			case "rejected":
+				badgeApply.rejectBadge();
+				break;
+			case "approved":
+				badgeApply.approveBadge();
+				user.badgeApproved();
+				break;
+			default:
+				throw new Exception400(process, "잘못된 요청값입니다.");
+		}
+	}
 
-    private void validateUserStatus(User user) {
-        if (user.getStatus() != UserStatus.NORMAL) {
-            throw new Exception404("해당 회원의 뱃지를 처리할 수 없습니다.");
-        }
-    }
+	private void validateUserStatus(User user) {
+		if (user.getStatus() != UserStatus.NORMAL) {
+			throw new Exception404("해당 회원의 뱃지를 처리할 수 없습니다.");
+		}
+	}
 
-    @Transactional(readOnly = true)
-    public MemberListResp getMemberList(String nickName, Flag badge, UserStatus userStatus, Pageable pageable) {
-        List<User> memberList = userRepository.findAllByFlagAndStatus(nickName, badge, userStatus, pageable);
+	@Transactional(readOnly = true)
+	public MemberListResp getMemberList(String nickName, Flag badge, UserStatus userStatus, Pageable pageable) {
+		List<User> memberList = userRepository.findAllByFlagAndStatus(nickName, badge, userStatus, pageable);
 
-        List<MemberListResp.memberDTO> memberDTOList = new ArrayList<>();
+		List<MemberListResp.memberDTO> memberDTOList = new ArrayList<>();
 
-        for (User member : memberList) {
-            Long feedCount = feedRepository.countByUserAndStatus(member, ContentStatus.NORMAL);
-            Long replyCount = replyRepository.countByUserAndStatus(member, ContentStatus.NORMAL);
+		for (User member : memberList) {
+			Long feedCount = feedRepository.countByUserAndStatus(member, ContentStatus.NORMAL);
+			Long replyCount = replyRepository.countByUserAndStatus(member, ContentStatus.NORMAL);
 
-            long approveCount = reportRepository.countProcessedByStatus(member, ProcessedStatus.APPROVED);
+			long approveCount = reportRepository.countProcessedByStatus(member, ProcessedStatus.APPROVED);
 
-            MemberListResp.memberDTO memberDTO = new MemberListResp.memberDTO(member, feedCount, replyCount, approveCount);
-            memberDTOList.add(memberDTO);
-        }
+			MemberListResp.memberDTO memberDTO = new MemberListResp.memberDTO(member, feedCount, replyCount,
+				approveCount);
+			memberDTOList.add(memberDTO);
+		}
 
-        return new MemberListResp(memberDTOList);
-    }
+		return new MemberListResp(memberDTOList);
+	}
 
-    @Transactional
-    public void blockProcessed(BlockReq request) {
-        User user = userRepository.findByIdAndStatus(request.getUserId(), UserStatus.NORMAL)
-                .orElseThrow(() -> new Exception404("해당 회원이 존재 하지 않습니다."));
+	@Transactional
+	public void blockProcessed(BlockProcessedParam parameter) {
+		User user = userRepository.findByIdAndStatus(parameter.getUserId(), UserStatus.NORMAL)
+			.orElseThrow(() -> new Exception404("해당 회원이 존재 하지 않습니다."));
 
-        Long feedCount = feedRepository.countByUserAndStatus(user, ContentStatus.NORMAL);
-        Long replyCount = replyRepository.countByUserAndStatus(user, ContentStatus.NORMAL);
+		Long feedCount = feedRepository.countByUserAndStatus(user, ContentStatus.NORMAL);
+		Long replyCount = replyRepository.countByUserAndStatus(user, ContentStatus.NORMAL);
 
-        BlockUser blockUser = BlockUser.createBlock(user, request.getReason(), feedCount, replyCount);
-        blockUserRepository.save(blockUser);
+		BlockUser blockUser = BlockUser.createBlock(user, parameter.getReason(), feedCount, replyCount);
+		blockUserRepository.save(blockUser);
 
-        user.block();
+		user.block();
 
-        List<Feed> feedList = feedRepository.findByUserIdAndStatus(user.getId(), ContentStatus.NORMAL);
-        feedList.forEach(Feed::deleteFeed);
+		List<Feed> feedList = feedRepository.findByUserIdAndStatus(user.getId(), ContentStatus.NORMAL);
+		feedList.forEach(Feed::deleteFeed);
 
-        List<Reply> replyList = replyRepository.findByUserIdAndStatus(user.getId(), ContentStatus.NORMAL);
-        replyList.forEach(Reply::deleteReply);
+		List<Reply> replyList = replyRepository.findByUserIdAndStatus(user.getId(), ContentStatus.NORMAL);
+		replyList.forEach(Reply::deleteReply);
 
-        sendBlockProcessedMail(user);
-    }
+		sendBlockProcessedMail(user);
+	}
 
-    private void sendBlockProcessedMail(User blockUser) {
-        mailService.sendEmail(blockUser.getEmail(), "[Foodielog] 영구 차단 내역", "님 차단 ㅅㄱ");
-    }
+	private void sendBlockProcessedMail(User blockUser) {
+		mailService.sendEmail(blockUser.getEmail(), "[Foodielog] 영구 차단 내역", "님 차단 ㅅㄱ");
+	}
 }
